Move review type to core module
