+++ core/modules/review/features/addReview.ts
@@ -6,7 +6,7 @@
  *
  * Part of [Review API Module](https://github.com/DivanteLtd/vue-storefront/tree/master/doc/api-modules)
  */
-import Review from '@vue-storefront/store/types/review/Review'
+import Review from '@vue-storefront/core/modules/review/types/Review'
 export const addReview = {
   methods: {
     addReview (review: Review) {
