+++ core/store/modules/review/actions.ts
@@ -9,7 +9,7 @@ import * as types from '../../mutation-types'
 import i18n from '@vue-storefront/i18n'
 import rootStore from "core/store"
 import { ValidationError } from "core/store/lib/exceptions";
-import Review from '@vue-storefront/store/types/review/Review'
+import Review from '@vue-storefront/core/modules/review/types/Review'
 const Ajv = require('ajv') // json validator
 
 const actions: ActionTree<ReviewState, RootState> = {
